level 2 part 2 correction according to the comments in the review. changed packages, moved data classes  and AppModule to separate folders.
