+++ app/src/main/java/com/project/myapp/MyApp.kt
@@ -3,6 +3,5 @@ package com.project.myapp
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
-
 @HiltAndroidApp
 class MyApp : Application()
