+++ app/src/main/java/com/project/myapp/data/contacts/UsersRepository.kt
@@ -1,4 +1,4 @@
-package com.project.myapp.screens.contacts
+package com.project.myapp.data.contacts
 
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
