+++ app/src/main/java/com/project/myapp/data/datastore/DataStore.kt
@@ -1,4 +1,4 @@
-package com.project.myapp
+package com.project.myapp.data.datastore
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
