+++ app/src/main/java/com/project/myapp/screens/auth/AuthActivity.kt
@@ -6,7 +6,7 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.widget.doOnTextChanged
 import androidx.lifecycle.lifecycleScope
-import com.project.myapp.KeysHolder.USER_NAME_KEY
+import com.project.myapp.data.datastore.KeysHolder.USER_NAME_KEY
 import com.project.myapp.R
 import com.project.myapp.databinding.ActivityAuthBinding
 import com.project.myapp.ext.componentactivity.enableEdgeToEdgeGrayStatusBar
