+++ app/src/main/java/com/project/myapp/screens/contacts/ContactViewModel.kt
@@ -1,6 +1,8 @@
 package com.project.myapp.screens.contacts
 
 import androidx.lifecycle.ViewModel
+import com.project.myapp.data.contacts.User
+import com.project.myapp.data.contacts.UsersRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
