+++ app/src/main/java/com/project/myapp/screens/splashScreen/SplashViewModel.kt
@@ -2,7 +2,7 @@ package com.project.myapp.screens.splashScreen
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.project.myapp.DataStore
+import com.project.myapp.data.datastore.DataStore
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
